Is this still needed?
GitlabCI with a Github hosted repository will actually have its
git remote origin point to gitlab.com and not to github.com.
The reason is that GitlabCI maintains a mirror of the Github Repo
and that the git remote in the CI points to the mirror rather than
to the actual git repository on Github.

I'm not sure if I see a need for maintain this #validates_as_ci?
for RequestSources...